+++ tests/testthat/test-env.R
@@ -216,3 +216,38 @@ test_that("env_tail() detects sentinel", {
   descendant <- child_env(child_env(child_env(env)))
   expect_identical(env_tail(descendant, sentinel), env)
 })
+
+test_that("env_get_list() retrieves multiple bindings", {
+  env <- env(foo = 1L, bar = 2L)
+  expect_identical(env_get_list(env, c("foo", "bar")), list(foo = 1L, bar =2L))
+
+  baz <- 0L
+  expect_error(env_get_list(env, "baz"), "'baz' not found")
+  expect_identical(env_get_list(env, c("foo", "baz"), inherit = TRUE), list(foo = 1L, baz =0L))
+})
+
+test_that("scoped_bindings binds temporarily", {
+  env <- env(foo = "foo", bar = "bar")
+
+  local({
+    old <- scoped_bindings(.env = env,
+      foo = "FOO",
+      bar = "BAR",
+      baz = "BAZ"
+    )
+    expect_identical(old, list(foo = "foo", bar = "bar"))
+    temp_bindings <- env_get_list(env, c("foo", "bar", "baz"))
+    expect_identical(temp_bindings, list(foo = "FOO", bar = "BAR", baz = "BAZ"))
+  })
+
+  bindings <- env_get_list(env, c("foo", "bar"))
+  expect_identical(bindings, list(foo = "foo", bar = "bar"))
+  expect_false(env_has(env, "baz"))
+})
+
+test_that("with_bindings() evaluates with temporary bindings", {
+  foo <- "foo"
+  baz <- "baz"
+  expect_identical(with_bindings(paste(foo, baz), foo = "FOO"), "FOO baz")
+  expect_identical(foo, "foo")
+})